Separate JournalOutput, also to avoid cycles
===============================================================================

Cherry picked from Audacity commit 00ca41abc4edbd7562c64a4d80e5a5edcb859f1b.
Modified for Saucedacity.
