+++ src/Journal.h
@@ -35,10 +35,6 @@ namespace Journal
    //\return whether successful
    bool SetRecordEnabled(bool value);
 
-   //\brief Whether actually recording.
-   // IsRecording() && IsReplaying() is possible
-   bool IsRecording();
-
    //\brief Whether actually replaying.
    // IsRecording() && IsReplaying() is possible
    bool IsReplaying();
@@ -63,16 +59,6 @@ namespace Journal
    // conditions are encountered.
    // Returns true if any command was dispatched
    bool Dispatch();
-
-   //\brief write the strings to the output journal, if recording
-   // None of them may contain newlines
-   void Output( const wxString &string );
-   void Output( const wxArrayString &strings );
-   void Output( std::initializer_list< const wxString > strings );
-
-   //\brief if recording, emit a comment in the output journal that will have
-   // no effect on playback
-   void Comment( const wxString &string );
 
    //\brief If recording, output the strings; if playing back, require
    // identical strings.  None of them may contain newlines
