+++ src/commands/CommandManager.cpp
@@ -91,6 +91,7 @@ CommandManager.  It holds the callback for one command.
 #include <wx/tokenzr.h>
 
 #include "../Journal.h"
+#include "../JournalOutput.h"
 #include "../JournalRegistry.h"
 #include "../Menus.h"
